- Renamed Architecture.Ddd.Services.Ensure.AggregateDoesNotConflict to Architecture.Ddd.Services.Ensure.DoesNotConflict (**Breaking Change**).
