+++ src/MooVC.Architecture/Ddd/Specifications/Ensure.Satisifies.cs
